type change

+++ src/main/java/sbt/dao/repository/ReceiptRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import sbt.dao.model.Receipt;
 import sbt.data.LoginInfo;
 
-public interface ReceiptRepository extends CrudRepository<Receipt, Integer> {
+public interface ReceiptRepository extends CrudRepository<Receipt, Long> {
 }
